arbor-core/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.9.2
nats-py==2.9.0
redis==5.2.0
asyncpg==0.30.0
httpx==0.27.2
zeroconf==0.131.0
python-dotenv==1.0.1
